Resize the date array once it's full of dates.
Prior functionality stopped adding dates to the dates array once it was
full. The array now dynamically resizes itself whenever it gets full of
dates.

An initialize method was added to initialize the first date array. This
method must be called before the menu is presented to the user and
before any dates are worked with.